fix: 로또 숫자를 `Long`이 아닌 `Integer`로 수정
과제에서 `Integer` 형태로 `Lotto`에 저장함에 따라 맞춰서 `Integer`로 수정함.

+++ src/main/java/lotto/InputView.java
@@ -6,10 +6,10 @@ import java.util.Arrays;
 import java.util.List;
 
 public class InputView {
-    public static long getAmountFromUser() {
+    public static int getAmountFromUser() {
         OutputView.promptForAmount();
         String input = readLine();
 
-        return StringParser.parseLong(input);
+        return StringParser.parseInt(input);
     }
 }
