+++ src/main/java/lotto/StringParser.java
@@ -1,9 +1,9 @@
 package lotto;
 
 public class StringParser {
-    public static long parseLong(String input) {
+    public static int parseInt(String input) {
         try {
-            return Long.parseLong(input);
+            return Integer.parseInt(input);
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("[ERROR] 숫자를 입력해 주세요.");
         }
